[93] Reorganize existing diagram modules to share common behaviors
 - Definition node description
 - Usage node description
 - Package node description
 - Fake node description
 - Tool services
 - Create services

Bug: https://github.com/eclipse-syson/syson/issues/93
