+++ backend/views/syson-diagram-actionflow-view/src/main/java/org/eclipse/syson/diagram/actionflow/view/nodes/FakeNodeDescriptionProvider.java
@@ -12,17 +12,15 @@
  *******************************************************************************/
 package org.eclipse.syson.diagram.actionflow.view.nodes;
 
+import java.util.ArrayList;
+import java.util.List;
+
 import org.eclipse.sirius.components.view.builder.IViewDiagramElementFinder;
 import org.eclipse.sirius.components.view.builder.providers.IColorProvider;
-import org.eclipse.sirius.components.view.diagram.DiagramDescription;
 import org.eclipse.sirius.components.view.diagram.NodeDescription;
-import org.eclipse.sirius.components.view.diagram.NodeStyleDescription;
-import org.eclipse.sirius.components.view.diagram.SynchronizationPolicy;
 import org.eclipse.syson.diagram.actionflow.view.AFVDescriptionNameGenerator;
 import org.eclipse.syson.diagram.actionflow.view.ActionFlowViewDiagramDescriptionProvider;
-import org.eclipse.syson.diagram.common.view.nodes.AbstractNodeDescriptionProvider;
-import org.eclipse.syson.sysml.SysmlPackage;
-import org.eclipse.syson.util.SysMLMetamodelHelper;
+import org.eclipse.syson.diagram.common.view.nodes.AbstractFakeNodeDescriptionProvider;
 
 /**
  * Fake Node Description allowing to store other Node Descriptions that will be reused by other Node Descriptions.
@@ -30,54 +28,25 @@ import org.eclipse.syson.util.SysMLMetamodelHelper;
  *
  * @author arichard
  */
-public class FakeNodeDescriptionProvider extends AbstractNodeDescriptionProvider {
-
-    public static final String NAME = "AFV Node Fake";
+public class FakeNodeDescriptionProvider extends AbstractFakeNodeDescriptionProvider {
 
     public FakeNodeDescriptionProvider(IColorProvider colorProvider) {
         super(colorProvider);
     }
 
     @Override
-    public NodeDescription create() {
-        String domainType = SysMLMetamodelHelper.buildQualifiedName(SysmlPackage.eINSTANCE.getElement());
-        return this.diagramBuilderHelper.newNodeDescription()
-                .defaultHeightExpression("0")
-                .defaultWidthExpression("0")
-                .domainType(domainType)
-                .labelExpression("")
-                .name(NAME)
-                .semanticCandidatesExpression("")
-                .style(this.createFakeNodeStyle())
-                .userResizable(true)
-                .synchronizationPolicy(SynchronizationPolicy.UNSYNCHRONIZED)
-                .build();
+    protected String getName() {
+        return "AFV Node Fake";
     }
 
     @Override
-    public void link(DiagramDescription diagramDescription, IViewDiagramElementFinder cache) {
+    protected List<NodeDescription> getChildrenDescription(IViewDiagramElementFinder cache) {
         var nameGenerator = new AFVDescriptionNameGenerator();
-        var optFakeNodeDescription = cache.getNodeDescription(NAME);
-        NodeDescription nodeDescription = optFakeNodeDescription.get();
-        diagramDescription.getNodeDescriptions().add(nodeDescription);
+        var childrenNodes = new ArrayList<NodeDescription>();
 
         ActionFlowViewDiagramDescriptionProvider.COMPARTMENTS_WITH_LIST_ITEMS.forEach((type, listItems) -> {
-            listItems.forEach(eReference -> {
-                var optNodeDescription = cache.getNodeDescription(nameGenerator.getCompartmentName(type, eReference));
-                nodeDescription.getChildrenDescriptions().add(optNodeDescription.get());
-            });
+            listItems.forEach(eReference -> cache.getNodeDescription(nameGenerator.getCompartmentName(type, eReference)).ifPresent(childrenNodes::add));
         });
-    }
-
-    protected NodeStyleDescription createFakeNodeStyle() {
-        return this.diagramBuilderHelper.newRectangularNodeStyleDescription()
-                .borderColor(this.colorProvider.getColor("transparent"))
-                .borderRadius(0)
-                .color(this.colorProvider.getColor("transparent"))
-                .displayHeaderSeparator(false)
-                .labelColor(this.colorProvider.getColor("transparent"))
-                .showIcon(false)
-                .withHeader(false)
-                .build();
+        return childrenNodes;
     }
 }
