+++ backend/views/syson-diagram-actionflow-view/src/main/java/org/eclipse/syson/diagram/actionflow/view/nodes/PackageNodeDescriptionProvider.java
@@ -15,239 +15,59 @@ package org.eclipse.syson.diagram.actionflow.view.nodes;
 import java.util.ArrayList;
 import java.util.List;
 
-import org.eclipse.emf.ecore.EClass;
 import org.eclipse.sirius.components.view.builder.IViewDiagramElementFinder;
-import org.eclipse.sirius.components.view.builder.generated.FreeFormLayoutStrategyDescriptionBuilder;
-import org.eclipse.sirius.components.view.builder.generated.NodeToolSectionBuilder;
 import org.eclipse.sirius.components.view.builder.providers.IColorProvider;
-import org.eclipse.sirius.components.view.diagram.DiagramDescription;
-import org.eclipse.sirius.components.view.diagram.DropNodeTool;
-import org.eclipse.sirius.components.view.diagram.EdgeTool;
-import org.eclipse.sirius.components.view.diagram.NodeContainmentKind;
 import org.eclipse.sirius.components.view.diagram.NodeDescription;
-import org.eclipse.sirius.components.view.diagram.NodePalette;
-import org.eclipse.sirius.components.view.diagram.NodeStyleDescription;
-import org.eclipse.sirius.components.view.diagram.NodeTool;
-import org.eclipse.sirius.components.view.diagram.NodeToolSection;
-import org.eclipse.sirius.components.view.diagram.SynchronizationPolicy;
 import org.eclipse.syson.diagram.actionflow.view.AFVDescriptionNameGenerator;
 import org.eclipse.syson.diagram.actionflow.view.ActionFlowViewDiagramDescriptionProvider;
-import org.eclipse.syson.diagram.common.view.nodes.AbstractNodeDescriptionProvider;
-import org.eclipse.syson.diagram.common.view.services.ViewEdgeToolSwitch;
+import org.eclipse.syson.diagram.common.view.nodes.AbstractPackageNodeDescriptionProvider;
+import org.eclipse.syson.diagram.common.view.tools.ToolSectionDescription;
 import org.eclipse.syson.sysml.SysmlPackage;
-import org.eclipse.syson.sysmlcustomnodes.SysMLCustomnodesFactory;
-import org.eclipse.syson.sysmlcustomnodes.SysMLPackageNodeStyleDescription;
-import org.eclipse.syson.util.AQLConstants;
-import org.eclipse.syson.util.SysMLMetamodelHelper;
-import org.eclipse.syson.util.ViewConstants;
 
 /**
- * Used to create the package node description in Action Flow diagram.
+ * Used to create the package node description in Action Flow View diagram.
  *
  * @author arichard
  */
-public class PackageNodeDescriptionProvider extends AbstractNodeDescriptionProvider {
-
-    private final AFVDescriptionNameGenerator nameGenerator = new AFVDescriptionNameGenerator();
+public class PackageNodeDescriptionProvider extends AbstractPackageNodeDescriptionProvider {
 
     public PackageNodeDescriptionProvider(IColorProvider colorProvider) {
-        super(colorProvider);
+        super(colorProvider, new AFVDescriptionNameGenerator());
     }
 
     @Override
-    public NodeDescription create() {
-        String domainType = SysMLMetamodelHelper.buildQualifiedName(SysmlPackage.eINSTANCE.getPackage());
-        return this.diagramBuilderHelper.newNodeDescription()
-                .collapsible(true)
-                .childrenLayoutStrategy(new FreeFormLayoutStrategyDescriptionBuilder().build())
-                .defaultHeightExpression("300")
-                .defaultWidthExpression("300")
-                .domainType(domainType)
-                .labelExpression("aql:self.getContainerLabel()")
-                .name(this.nameGenerator.getNodeName(SysmlPackage.eINSTANCE.getPackage()))
-                .semanticCandidatesExpression("aql:self.getAllReachable(" + domainType + ")")
-                .style(this.createPackageNodeStyle())
-                .userResizable(true)
-                .synchronizationPolicy(SynchronizationPolicy.UNSYNCHRONIZED)
-                .build();
+    protected List<NodeDescription> getReusedChildren(IViewDiagramElementFinder cache) {
+        var reusedChildren = new ArrayList<NodeDescription>();
+
+        ActionFlowViewDiagramDescriptionProvider.DEFINITIONS.forEach(definition -> cache.getNodeDescription(this.nameGenerator.getNodeName(definition)).ifPresent(reusedChildren::add));
+        ActionFlowViewDiagramDescriptionProvider.USAGES.forEach(usage -> cache.getNodeDescription(this.nameGenerator.getNodeName(usage)).ifPresent(reusedChildren::add));
+        cache.getNodeDescription(this.nameGenerator.getNodeName(SysmlPackage.eINSTANCE.getPackage())).ifPresent(reusedChildren::add);
+        return reusedChildren;
     }
 
     @Override
-    public void link(DiagramDescription diagramDescription, IViewDiagramElementFinder cache) {
-        var optPackageNodeDescription = cache.getNodeDescription(this.nameGenerator.getNodeName(SysmlPackage.eINSTANCE.getPackage()));
-        NodeDescription packageNodeDescription = optPackageNodeDescription.get();
-        diagramDescription.getNodeDescriptions().add(packageNodeDescription);
+    protected List<NodeDescription> getDroppableNodes(IViewDiagramElementFinder cache) {
+        var droppableNodes = new ArrayList<NodeDescription>();
 
-        var allTargetNodeDescriptions = new ArrayList<NodeDescription>();
+        ActionFlowViewDiagramDescriptionProvider.DEFINITIONS.forEach(definition -> cache.getNodeDescription(this.nameGenerator.getNodeName(definition)).ifPresent(droppableNodes::add));
+        ActionFlowViewDiagramDescriptionProvider.USAGES.forEach(usage -> cache.getNodeDescription(this.nameGenerator.getNodeName(usage)).ifPresent(droppableNodes::add));
+        cache.getNodeDescription(this.nameGenerator.getNodeName(SysmlPackage.eINSTANCE.getPackage())).ifPresent(droppableNodes::add);
 
-        ActionFlowViewDiagramDescriptionProvider.DEFINITIONS.forEach(definition -> {
-            var optNodeDescription = cache.getNodeDescription(this.nameGenerator.getNodeName(definition));
-            allTargetNodeDescriptions.add(optNodeDescription.get());
-            packageNodeDescription.getReusedChildNodeDescriptions().add(optNodeDescription.get());
-        });
-
-        ActionFlowViewDiagramDescriptionProvider.USAGES.forEach(usage -> {
-            var optNodeDescription = cache.getNodeDescription(this.nameGenerator.getNodeName(usage));
-            allTargetNodeDescriptions.add(optNodeDescription.get());
-            packageNodeDescription.getReusedChildNodeDescriptions().add(optNodeDescription.get());
-        });
-
-        allTargetNodeDescriptions.add(packageNodeDescription);
-        packageNodeDescription.getReusedChildNodeDescriptions().add(packageNodeDescription);
-        packageNodeDescription.setPalette(this.createNodePalette(packageNodeDescription, cache, allTargetNodeDescriptions));
+        return droppableNodes;
     }
 
-    protected NodeStyleDescription createPackageNodeStyle() {
-        SysMLPackageNodeStyleDescription nodeStyleDescription = SysMLCustomnodesFactory.eINSTANCE.createSysMLPackageNodeStyleDescription();
-        nodeStyleDescription.setBorderColor(this.colorProvider.getColor(ViewConstants.DEFAULT_BORDER_COLOR));
-        nodeStyleDescription.setBorderRadius(0);
-        nodeStyleDescription.setColor(this.colorProvider.getColor(ViewConstants.DEFAULT_BACKGROUND_COLOR));
-        nodeStyleDescription.setLabelColor(this.colorProvider.getColor(ViewConstants.DEFAULT_LABEL_COLOR));
-        nodeStyleDescription.setShowIcon(true);
-        return nodeStyleDescription;
+    @Override
+    protected List<NodeDescription> getAllNodeDescriptions(IViewDiagramElementFinder cache) {
+        var allNodes = new ArrayList<NodeDescription>();
+
+        ActionFlowViewDiagramDescriptionProvider.DEFINITIONS.forEach(definition -> cache.getNodeDescription(this.nameGenerator.getNodeName(definition)).ifPresent(allNodes::add));
+        ActionFlowViewDiagramDescriptionProvider.USAGES.forEach(usage -> cache.getNodeDescription(this.nameGenerator.getNodeName(usage)).ifPresent(allNodes::add));
+        cache.getNodeDescription(this.nameGenerator.getNodeName(SysmlPackage.eINSTANCE.getPackage())).ifPresent(allNodes::add);
+        return allNodes;
     }
 
-    private NodePalette createNodePalette(NodeDescription nodeDescription, IViewDiagramElementFinder cache, List<NodeDescription> allNodeDescriptions) {
-        var changeContext = this.viewBuilderHelper.newChangeContext()
-                .expression("aql:self.deleteFromModel()");
-
-        var deleteTool = this.diagramBuilderHelper.newDeleteTool()
-                .name("Delete from Model")
-                .body(changeContext.build());
-
-        var callEditService = this.viewBuilderHelper.newChangeContext()
-                .expression(AQLConstants.AQL_SELF + ".directEdit(newLabel)");
-
-        var editTool = this.diagramBuilderHelper.newLabelEditTool()
-                .name("Edit")
-                .initialDirectEditLabelExpression(AQLConstants.AQL_SELF + ".getDefaultInitialDirectEditLabel()")
-                .body(callEditService.build());
-
-        var edgeTools = new ArrayList<EdgeTool>();
-        edgeTools.addAll(this.getEdgeTools(nodeDescription, allNodeDescriptions));
-
-        return this.diagramBuilderHelper.newNodePalette()
-                .deleteTool(deleteTool.build())
-                .labelEditTool(editTool.build())
-                .dropNodeTool(this.createDropFromDiagramTool(cache))
-                .toolSections(this.createToolSections(cache))
-                .edgeTools(edgeTools.toArray(EdgeTool[]::new))
-                .build();
-    }
-
-    private List<EdgeTool> getEdgeTools(NodeDescription nodeDescription, List<NodeDescription> allNodeDescriptions) {
-        ViewEdgeToolSwitch edgeToolSwitch = new ViewEdgeToolSwitch(nodeDescription, allNodeDescriptions, this.nameGenerator);
-        edgeToolSwitch.doSwitch(SysmlPackage.eINSTANCE.getPackage());
-        return edgeToolSwitch.getEdgeTools();
-    }
-
-    private DropNodeTool createDropFromDiagramTool(IViewDiagramElementFinder cache) {
-        var acceptedNodeTypes = new ArrayList<NodeDescription>();
-
-        ActionFlowViewDiagramDescriptionProvider.DEFINITIONS.forEach(definition -> {
-            var optNodeDescription = cache.getNodeDescription(this.nameGenerator.getNodeName(definition));
-            acceptedNodeTypes.add(optNodeDescription.get());
-        });
-
-        ActionFlowViewDiagramDescriptionProvider.USAGES.forEach(usage -> {
-            var optNodeDescription = cache.getNodeDescription(this.nameGenerator.getNodeName(usage));
-            acceptedNodeTypes.add(optNodeDescription.get());
-        });
-
-        var optPackageNodeDescription = cache.getNodeDescription(this.nameGenerator.getNodeName(SysmlPackage.eINSTANCE.getPackage()));
-        acceptedNodeTypes.add(optPackageNodeDescription.get());
-
-        var dropElementFromDiagram = this.viewBuilderHelper.newChangeContext()
-                .expression("aql:droppedElement.dropElementFromDiagram(droppedNode, targetElement, targetNode, editingContext, diagramContext, convertedNodes)");
-
-        return this.diagramBuilderHelper.newDropNodeTool()
-                .name("Drop from Diagram")
-                .acceptedNodeTypes(acceptedNodeTypes.toArray(NodeDescription[]::new))
-                .body(dropElementFromDiagram.build())
-                .build();
-    }
-
-    private NodeTool createNodeTool(NodeDescription nodeDescription, EClass eClass) {
-        var setValue = this.viewBuilderHelper.newSetValue()
-                .featureName(SysmlPackage.eINSTANCE.getElement_DeclaredName().getName())
-                .valueExpression(eClass.getName());
-
-        var changeContextNewInstance = this.viewBuilderHelper.newChangeContext()
-                .expression("aql:newInstance")
-                .children(setValue.build());
-
-        var createEClassInstance = this.viewBuilderHelper.newCreateInstance()
-                .typeName(SysMLMetamodelHelper.buildQualifiedName(eClass))
-                .referenceName(SysmlPackage.eINSTANCE.getRelationship_OwnedRelatedElement().getName())
-                .variableName("newInstance")
-                .children(changeContextNewInstance.build());
-
-        var createView = this.diagramBuilderHelper.newCreateView()
-                .containmentKind(NodeContainmentKind.CHILD_NODE)
-                .elementDescription(nodeDescription)
-                .parentViewExpression("aql:selectedNode")
-                .semanticElementExpression("aql:newInstance")
-                .variableName("newInstanceView");
-
-        var changeContexMembership = this.viewBuilderHelper.newChangeContext()
-                .expression("aql:newOwningMembership")
-                .children(createEClassInstance.build(), createView.build());
-
-        var createMembership = this.viewBuilderHelper.newCreateInstance()
-                .typeName(SysMLMetamodelHelper.buildQualifiedName(SysmlPackage.eINSTANCE.getOwningMembership()))
-                .referenceName(SysmlPackage.eINSTANCE.getElement_OwnedRelationship().getName())
-                .variableName("newOwningMembership")
-                .children(changeContexMembership.build());
-
-        return this.diagramBuilderHelper.newNodeTool()
-                .name(this.nameGenerator.getCreationToolName(eClass))
-                .iconURLsExpression("/icons/full/obj16/" + eClass.getName() + ".svg")
-                .body(createMembership.build())
-                .build();
-    }
-
-    private NodeToolSection addElementsToolSection() {
-        return this.diagramBuilderHelper.newNodeToolSection()
-                .name("Add")
-                .nodeTools(this.addExistingElementsTool())
-                .build();
-    }
-
-    private NodeTool addExistingElementsTool() {
-        var addExistingelements = this.viewBuilderHelper.newChangeContext()
-                .expression("aql:self.addExistingElements(editingContext, diagramContext, selectedNode, convertedNodes)");
-
-        return this.diagramBuilderHelper.newNodeTool()
-                .name("Add existing elements")
-                .iconURLsExpression("/icons/AddExistingElements.svg")
-                .body(addExistingelements.build())
-                .build();
-    }
-
-    private NodeToolSection[] createToolSections(IViewDiagramElementFinder cache) {
-        var sections = new ArrayList<NodeToolSection>();
-
-        ActionFlowViewDiagramDescriptionProvider.TOOL_SECTIONS.forEach(sectionTool -> {
-            NodeToolSectionBuilder sectionBuilder = this.diagramBuilderHelper.newNodeToolSection()
-                    .name(sectionTool.name())
-                    .nodeTools(this.createElementsOfToolSection(cache, sectionTool.elements()));
-            sections.add(sectionBuilder.build());
-        });
-
-        sections.add(this.addElementsToolSection());
-
-        return sections.toArray(NodeToolSection[]::new);
-    }
-
-    private NodeTool[] createElementsOfToolSection(IViewDiagramElementFinder cache, List<EClass> elements) {
-        var nodeTools = new ArrayList<NodeTool>();
-
-        elements.forEach(definition -> {
-            nodeTools.add(this.createNodeTool(cache.getNodeDescription(this.nameGenerator.getNodeName(definition)).get(), definition));
-        });
-
-        nodeTools.sort((nt1, nt2) -> nt1.getName().compareTo(nt2.getName()));
-
-        return nodeTools.toArray(NodeTool[]::new);
+    @Override
+    protected List<ToolSectionDescription> getToolSections() {
+        return ActionFlowViewDiagramDescriptionProvider.TOOL_SECTIONS;
     }
 }
