+++ backend/views/syson-diagram-actionflow-view/src/main/java/org/eclipse/syson/diagram/actionflow/view/services/ActionFlowViewCreateService.java
@@ -12,18 +12,13 @@
  *******************************************************************************/
 package org.eclipse.syson.diagram.actionflow.view.services;
 
-import java.util.List;
-import java.util.Objects;
-
 import org.eclipse.sirius.components.collaborative.diagrams.api.IDiagramContext;
 import org.eclipse.sirius.components.core.api.IEditingContext;
 import org.eclipse.sirius.components.diagrams.Diagram;
-import org.eclipse.sirius.components.diagrams.Node;
-import org.eclipse.sirius.components.diagrams.ViewCreationRequest;
 import org.eclipse.sirius.components.view.emf.IViewRepresentationDescriptionSearchService;
 import org.eclipse.syson.diagram.actionflow.view.ActionFlowViewDiagramDescriptionProvider;
 import org.eclipse.syson.diagram.actionflow.view.nodes.ActionFlowViewEmptyDiagramNodeDescriptionProvider;
-import org.eclipse.syson.services.ElementInitializerSwitch;
+import org.eclipse.syson.diagram.common.view.services.ViewCreateService;
 import org.eclipse.syson.sysml.Element;
 
 /**
@@ -31,27 +26,10 @@ import org.eclipse.syson.sysml.Element;
  *
  * @author arichard
  */
-public class ActionFlowViewCreateService {
-
-    private final IViewRepresentationDescriptionSearchService viewRepresentationDescriptionSearchService;
-
-    private final ElementInitializerSwitch elementInitializerSwitch;
+public class ActionFlowViewCreateService extends ViewCreateService {
 
     public ActionFlowViewCreateService(IViewRepresentationDescriptionSearchService viewRepresentationDescriptionSearchService) {
-        this.viewRepresentationDescriptionSearchService = Objects.requireNonNull(viewRepresentationDescriptionSearchService);
-        this.elementInitializerSwitch = new ElementInitializerSwitch();
-    }
-
-    /**
-     * Call the {@link ElementInitializerSwitch} on the given {@link Element}. Allows to set various
-     * attributes/references.
-     *
-     * @param element
-     *            the given {@link Element}.
-     * @return the given {@link Element}.
-     */
-    public Element elementInitializer(Element element) {
-        return this.elementInitializerSwitch.doSwitch(element);
+        super(viewRepresentationDescriptionSearchService);
     }
 
     /**
@@ -68,20 +46,6 @@ public class ActionFlowViewCreateService {
      * @return the given {@link Element} if the diagram is empty, <code>null</code> otherwise.
      */
     public Element getDiagramEmptyCandidate(Element element, IEditingContext editingContext, IDiagramContext diagramContext, Diagram previousDiagram) {
-        boolean emptyDiagram = false;
-        if (previousDiagram != null && diagramContext != null) {
-            List<Node> previousNodes = previousDiagram.getNodes();
-            List<ViewCreationRequest> viewCreationRequests = diagramContext.getViewCreationRequests();
-            if (viewCreationRequests.isEmpty() && (previousNodes.isEmpty() || previousNodes.stream().anyMatch(node -> this.viewRepresentationDescriptionSearchService
-                    .findViewNodeDescriptionById(editingContext, node.getDescriptionId()).stream().anyMatch(nd -> ActionFlowViewEmptyDiagramNodeDescriptionProvider.NAME.equals(nd.getName()))))) {
-                emptyDiagram = true;
-            }
-        } else {
-            emptyDiagram = true;
-        }
-        if (emptyDiagram) {
-            return element;
-        }
-        return null;
+        return this.getDiagramEmptyCandidate(element, editingContext, diagramContext, previousDiagram, ActionFlowViewEmptyDiagramNodeDescriptionProvider.NAME);
     }
 }
