+++ backend/views/syson-diagram-common-view/src/main/java/org/eclipse/syson/diagram/common/view/edges/AbstractDependencyEdgeDescriptionProvider.java
@@ -44,19 +44,22 @@ public abstract class AbstractDependencyEdgeDescriptionProvider extends Abstract
 
     /**
      * Implementers should provide the actual name of this {@link EdgeDescription}.
-     * @return
+     *
+     * @return the name of the edge description
      */
     protected abstract String getName();
 
     /**
-     * implementers should provide the list of {@link NodeDescription} that can be a source of this {@link EdgeDescription}.
+     * Implementers should provide the list of {@link NodeDescription} that can be a source of this {@link EdgeDescription}.
+     *
      * @param cache the cache used to retrieve node descriptions.
      * @return the list of {@link NodeDescription} that can be a source of this {@link EdgeDescription}.
      */
     protected abstract List<NodeDescription> getSourceNodes(IViewDiagramElementFinder cache);
 
     /**
-     * implementers should provide the list of {@link NodeDescription} that can be a target of this {@link EdgeDescription}.
+     * Implementers should provide the list of {@link NodeDescription} that can be a target of this {@link EdgeDescription}.
+     *
      * @param cache the cache used to retrieve node descriptions.
      * @return the list of {@link NodeDescription} that can be a target of this {@link EdgeDescription}.
      */
@@ -84,8 +87,8 @@ public abstract class AbstractDependencyEdgeDescriptionProvider extends Abstract
         EdgeDescription edgeDescription = optEdgeDescription.get();
         diagramDescription.getEdgeDescriptions().add(edgeDescription);
 
-        this.getSourceNodes(cache).forEach(edgeDescription.getSourceNodeDescriptions()::add);
-        this.getTargetNodes(cache).forEach(edgeDescription.getTargetNodeDescriptions()::add);
+        edgeDescription.getSourceNodeDescriptions().addAll(this.getSourceNodes(cache));
+        edgeDescription.getTargetNodeDescriptions().addAll(this.getTargetNodes(cache));
 
         edgeDescription.setPalette(this.createEdgePalette(List.of(this.createSourceReconnectTool(),
                 this.createTargetReconnectTool())));
