+++ backend/views/syson-diagram-common-view/src/main/java/org/eclipse/syson/diagram/common/view/nodes/AbstractDefinitionNodeDescriptionProvider.java
@@ -1,0 +1,154 @@
+/*******************************************************************************
+ * Copyright (c) 2024 Obeo.
+ * This program and the accompanying materials
+ * are made available under the terms of the Eclipse Public License v2.0
+ * which accompanies this distribution, and is available at
+ * https://www.eclipse.org/legal/epl-2.0/
+ *
+ * SPDX-License-Identifier: EPL-2.0
+ *
+ * Contributors:
+ *     Obeo - initial API and implementation
+ *******************************************************************************/
+package org.eclipse.syson.diagram.common.view.nodes;
+
+import java.util.ArrayList;
+import java.util.List;
+import java.util.Objects;
+
+import org.eclipse.emf.ecore.EClass;
+import org.eclipse.sirius.components.view.builder.IViewDiagramElementFinder;
+import org.eclipse.sirius.components.view.builder.providers.IColorProvider;
+import org.eclipse.sirius.components.view.diagram.DiagramDescription;
+import org.eclipse.sirius.components.view.diagram.EdgeTool;
+import org.eclipse.sirius.components.view.diagram.NodeDescription;
+import org.eclipse.sirius.components.view.diagram.NodePalette;
+import org.eclipse.sirius.components.view.diagram.NodeStyleDescription;
+import org.eclipse.sirius.components.view.diagram.NodeToolSection;
+import org.eclipse.sirius.components.view.diagram.SynchronizationPolicy;
+import org.eclipse.sirius.components.view.diagram.ToolSection;
+import org.eclipse.syson.diagram.common.view.services.ViewEdgeToolSwitch;
+import org.eclipse.syson.util.AQLConstants;
+import org.eclipse.syson.util.IDescriptionNameGenerator;
+import org.eclipse.syson.util.SysMLMetamodelHelper;
+import org.eclipse.syson.util.ViewConstants;
+
+/**
+ * Node description provider for all SysMLv2 Definitions elements.
+ *
+ * @author arichard
+ */
+public abstract class AbstractDefinitionNodeDescriptionProvider extends AbstractNodeDescriptionProvider {
+
+    protected final IDescriptionNameGenerator nameGenerator;
+
+    protected final EClass eClass;
+
+    public AbstractDefinitionNodeDescriptionProvider(EClass eClass, IColorProvider colorProvider, IDescriptionNameGenerator nameGenerator) {
+        super(colorProvider);
+        this.eClass = Objects.requireNonNull(eClass);
+        this.nameGenerator = Objects.requireNonNull(nameGenerator);
+    }
+
+    /**
+     * Implementers should provide the list of {@link NodeDescription} that is reused as child in that Definition node description.
+     *
+     * @param cache the cache used to retrieve node descriptions.
+     * @return the list of {@link NodeDescription} that are added as reused child.
+     */
+    protected abstract List<NodeDescription> getReusedChildren(IViewDiagramElementFinder cache);
+
+    /**
+     * Implementers should provide the list of all {@link NodeDescription} defined in the diagram.<br>
+     * This list is used to create edge tools associated to this Definition node description.
+     *
+     * @param cache the cache used to retrieve node descriptions.
+     * @return the list of all {@link NodeDescription} defined in the diagram.
+     */
+    protected abstract List<NodeDescription> getAllNodeDescriptions(IViewDiagramElementFinder cache);
+
+    /**
+     * Implementers should provide the list of {@link ToolSection} defined on the given {@link NodeDescription}.
+     *
+     * @param nodeDescription the actual Definition node description
+     * @param cache the cache used to retrieve node descriptions.
+     * @return the list of {@link ToolSection} of this Definition node.
+     */
+    protected abstract List<NodeToolSection> getToolSections(NodeDescription nodeDescription, IViewDiagramElementFinder cache);
+
+    @Override
+    public NodeDescription create() {
+        String domainType = SysMLMetamodelHelper.buildQualifiedName(this.eClass);
+        return this.diagramBuilderHelper.newNodeDescription()
+                .childrenLayoutStrategy(this.diagramBuilderHelper.newListLayoutStrategyDescription().areChildNodesDraggableExpression("false").build())
+                .collapsible(true)
+                .defaultHeightExpression(ViewConstants.DEFAULT_CONTAINER_NODE_HEIGHT)
+                .defaultWidthExpression(ViewConstants.DEFAULT_NODE_WIDTH)
+                .domainType(domainType)
+                .labelExpression("aql:self.getContainerLabel()")
+                .name(this.nameGenerator.getNodeName(this.eClass))
+                .semanticCandidatesExpression("aql:self.getAllReachable(" + domainType + ")")
+                .style(this.createDefinitionNodeStyle())
+                .userResizable(true)
+                .synchronizationPolicy(SynchronizationPolicy.UNSYNCHRONIZED)
+                .build();
+    }
+
+    @Override
+    public void link(DiagramDescription diagramDescription, IViewDiagramElementFinder cache) {
+        NodeDescription nodeDescription = cache.getNodeDescription(this.nameGenerator.getNodeName(this.eClass)).get();
+        diagramDescription.getNodeDescriptions().add(nodeDescription);
+
+        this.getReusedChildren(cache).forEach(nodeDescription.getReusedChildNodeDescriptions()::add);
+
+        nodeDescription.setPalette(this.createNodePalette(nodeDescription, cache));
+    }
+
+    private NodeStyleDescription createDefinitionNodeStyle() {
+        return this.diagramBuilderHelper.newRectangularNodeStyleDescription()
+                .borderColor(this.colorProvider.getColor(ViewConstants.DEFAULT_BORDER_COLOR))
+                .borderRadius(0)
+                .color(this.colorProvider.getColor(ViewConstants.DEFAULT_BACKGROUND_COLOR))
+                .displayHeaderSeparator(true)
+                .labelColor(this.colorProvider.getColor(ViewConstants.DEFAULT_LABEL_COLOR))
+                .showIcon(true)
+                .withHeader(true)
+                .build();
+    }
+
+    private NodePalette createNodePalette(NodeDescription nodeDescription, IViewDiagramElementFinder cache) {
+        var changeContext = this.viewBuilderHelper.newChangeContext()
+                .expression(AQLConstants.AQL_SELF + ".deleteFromModel()");
+
+        var deleteTool = this.diagramBuilderHelper.newDeleteTool()
+                .name("Delete from Model")
+                .body(changeContext.build());
+
+        var callEditService = this.viewBuilderHelper.newChangeContext()
+                .expression(AQLConstants.AQL_SELF + ".directEdit(newLabel)");
+
+        var editTool = this.diagramBuilderHelper.newLabelEditTool()
+                .name("Edit")
+                .initialDirectEditLabelExpression(AQLConstants.AQL_SELF + ".getDefaultInitialDirectEditLabel()")
+                .body(callEditService.build());
+
+        var edgeTools = new ArrayList<EdgeTool>();
+        edgeTools.addAll(this.getEdgeTools(nodeDescription, cache));
+
+        var toolSections = new ArrayList<NodeToolSection>();
+        toolSections.addAll(this.getToolSections(nodeDescription, cache));
+
+        return this.diagramBuilderHelper.newNodePalette()
+                .deleteTool(deleteTool.build())
+                .labelEditTool(editTool.build())
+                .edgeTools(edgeTools.toArray(EdgeTool[]::new))
+                .toolSections(toolSections.toArray(NodeToolSection[]::new))
+                .build();
+    }
+
+    private List<EdgeTool> getEdgeTools(NodeDescription nodeDescription, IViewDiagramElementFinder cache) {
+        ViewEdgeToolSwitch edgeToolSwitch = new ViewEdgeToolSwitch(nodeDescription, this.getAllNodeDescriptions(cache), this.nameGenerator);
+        edgeToolSwitch.doSwitch(this.eClass);
+        return edgeToolSwitch.getEdgeTools();
+    }
+}
