+++ backend/views/syson-diagram-general-view/src/main/java/org/eclipse/syson/diagram/general/view/GeneralViewJavaServiceProvider.java
@@ -17,9 +17,9 @@ import java.util.List;
 import org.eclipse.sirius.components.view.View;
 import org.eclipse.sirius.components.view.emf.IJavaServiceProvider;
 import org.eclipse.syson.diagram.common.view.services.ViewLabelService;
+import org.eclipse.syson.diagram.common.view.services.ViewToolService;
 import org.eclipse.syson.diagram.general.view.services.GeneralViewCreateService;
 import org.eclipse.syson.diagram.general.view.services.GeneralViewEdgeService;
-import org.eclipse.syson.diagram.general.view.services.GeneralViewToolService;
 import org.eclipse.syson.services.DeleteService;
 import org.eclipse.syson.services.UtilService;
 import org.springframework.context.annotation.Configuration;
@@ -43,7 +43,7 @@ public class GeneralViewJavaServiceProvider implements IJavaServiceProvider {
                     DeleteService.class,
                     GeneralViewEdgeService.class,
                     ViewLabelService.class,
-                    GeneralViewToolService.class,
+                    ViewToolService.class,
                     UtilService.class);
         }
         return List.of();
